+++ BudgetCoreDAO/Modifier.cs
@@ -15,8 +15,8 @@ namespace BudgetCoreDAO
         /// <summary>Наценка (ставка).</summary>
         public decimal Markup { get; set; }
         /// <summary>Скидка (сумма), в копейках.</summary>
-        public decimal DiscountSum { get; set; }
+        public int DiscountSum { get; set; }
         /// <summary>yНаценка (сумма), в копейках.</summary>
-        public decimal MarkupSum { get; set; }
+        public int MarkupSum { get; set; }
     }
 }
